modd .js to index and .css to style to match current master
